# docs/swagger.yaml
swagger: '2.0'
info:
  title: Predico - Restful-API
  description: |

    #### Description:
    Restful API for the INESC TEC Data Sharing / Barter Incentives Service.

    This API is crafted to enable secure data exchange and incentivized collaboration among multiple agents and a central server, where the execution of Data Sharing/Barter Incentives algorithms takes place. It robustly supports internal mechanisms for Data Sharing/Barter Incentives, addressing a diverse range of use cases.

    A prominent use case, characterized by a higher Technology Readiness Level (TRL), is the collaborative forecasting scenario.
    In this context, various agents can trade historical (or future) time-series data—such as power generation or weather measurements data—and acquire time-series forecasts improved by the collective pool of data available on the platform. Notably, agents can access these forecasts without directly tapping into internal datasets. The economic value of data shared by one agent is inherently tied to its utility for other participants that also buy collaborative forecasts from this service.

    ## Main functions:
    - **User management**:
        * Register new user
        * User verification / authentication
        * User password reset
        * Manage User Resources (i.e., bidding units in user portfolio) (create/update/delete)
        * Manage User IOTA wallet address reference (create/update/delete)

    - **Market management**:
...

...

...
  license:
...
  version: v1
host: predico02.inesctec.pt
schemes:
...
basePath: /api
consumes:
...
produces:
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
paths:
...
  /market/balance/:
    get:
      operationId: get_market_balance
      description: Method to get market balance for a given agent
      parameters:
        - name: balance__gte
          in: query
          description: Balance filter (greater or equal than).
          required: false
          type: string
        - name: balance__lte
          in: query
          description: Balance filter (lower or equal than).
          required: false
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 200
              data:
                description: Response data.
                type: object
          examples:
            application/json:
              code: 200
              data:
                - user: 1
                  balance: 0
                  total_deposit: 10000000
                  total_withdraw: 6000000
                  total_payment: 5000000
                  total_revenue: 1000000
                  updated_at: '2022-07-25T11:08:18.289370Z'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - market
    parameters: []
...
  /market/session-balance/:
    get:
      operationId: get_market_session_balance
      description: Method to get market balance of a given agent resource, in a specific market session
      parameters:
        - name: balance__gte
          in: query
          description: Balance filter (greater or equal than).
          required: false
          type: string
        - name: balance__lte
          in: query
          description: Balance filter (lower or equal than).
          required: false
          type: string
        - name: balance_by_resource
          in: query
          description: If true, returns balance by individual resource of each user in each session. Else, if false, returns the sum of balance per user in each session.
          required: false
          type: boolean
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 200
              data:
                description: Response data.
                type: object
          examples:
            application/json:
              code: 200
              data:
                - market_session: 1
                  user: 1
                  resource: 1
                  session_balance: 3000000
                  session_deposit: 3000000
                  session_payment: 0
                  session_revenue: 0
                  registered_at: '2022-10-20T11:40:22.364907Z'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - market
    post:
      operationId: post_market_session_balance
      description: '[AdminOnly] Method to insert balance of a given agent resource, in a specific market session'
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/MarketSessionBalanceCreate'
      responses:
        '201':
          description: Updated session balance for the user ID (<user_id>).
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - market
    parameters: []
...
  /market/transfer-out/:
    get:
      operationId: get_balance_transfer_out
      description: '[AdminOnly] Method to list historical balance transfers (i.e., from market wallet to agents wallet)'
      parameters:
        - name: is_solid
          in: query
          description: Filter transactions which are solid (confirmed) in IOTA Tangle
          required: false
          type: boolean
          default: false
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 200
              data:
                description: Response data.
                type: object
          examples:
            application/json:
              code: 200
              data: []
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - market
    post:
      operationId: post_balance_transfer_out
      description: '[AdminOnly] Method to register new balance transfer'
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/BalanceTransferOut'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/BalanceTransferOut'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - market
    put:
      operationId: put_balance_transfer_out
      description: '[AdminOnly] Method to update balance transfer'
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/BalanceTransferOut'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/BalanceTransferOut'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - market
    parameters: []
...
  /user/resource/:
    get:
      operationId: get_user_resources
      description: Method for agents to list their registered resources
      parameters:
        - name: resource
          in: query
          description: Filter by resource identifier
          required: false
          type: integer
        - name: resource_name
          in: query
          description: Filter by resource name
          required: false
          type: string
        - name: to_forecast
          in: query
          description: Return resources currently registered as 'to_forecast' in the market
          required: false
          type: boolean
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 200
              data:
                description: Response data.
                type: object
          examples:
            application/json:
              code: 200
              data:
                - id: 1
                  name: park-1
                  type: measurements
                  to_forecast: true
                  registered_at: '2022-10-20T11:34:57.739206Z'
                  user: 1
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - user
    post:
      operationId: post_user_resources
      description: Method for agents to register a new resource
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserResources'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/UserResources'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - user
    parameters: []
...
definitions:
...
  MarketSessionBalanceCreate:
    required:
      - user
      - market_session
      - resource
      - transaction_type
      - amount
    type: object
    properties:
      user:
        title: User
        type: integer
      market_session:
        title: Market session
        type: integer
      resource:
        title: Resource
        type: integer
      transaction_type:
        title: Transaction type
        type: string
        enum:
          - payment
          - revenue
          - transfer_in
          - transfer_out
      amount:
        title: Amount
        type: number
  BalanceTransferOut:
    required:
      - user_wallet_address
      - tangle_msg_id
      - user
    type: object
    properties:
      withdraw_transfer_id:
        title: Withdraw transfer id
        type: integer
        readOnly: true
      amount:
        title: Amount
        type: integer
        maximum: 2147483647
        minimum: -2147483648
      user_wallet_address:
        title: User wallet address
        type: string
        minLength: 1
      tangle_msg_id:
        title: Tangle msg id
        type: string
        minLength: 1
      is_solid:
        title: Is solid
        type: boolean
      registered_at:
        title: Registered at
        type: string
        format: date-time
        readOnly: true
      user:
        title: User
        type: integer
...